Added giant hunk of word2hangeul2.
I don't really like the design, since it's too
messy.  Also, it doesn't currently handle all
cases correctly, since I assumed that final
consonants were a superset of initial consonants,
which is not true.  For example, ㄸ is an initial
consonant, but not a final one.  This means that
people can't type 다따 in the current
implementation. I think this might be a good
candidate application to develop in a test-driven
style.