Added table of compatibility jamo to translator.
This looks a bit more ugly than the code for doing
the regular jamo, since regular jamo and
characters can both be computed with the same
offsets, but the compatibility jamo require their
own offset table.  I'm convinved that it's the
right thing to do, however, so I'm sticking with
it.